--- result_of_server/0408_gemini/ssh-remote+mqubit.skku.edu:.inetloc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>URL</key>
-	<string>vscode-remote://ssh-remote%2Bmqubit.skku.edu/home/jaeheekim/QC/LQN_gemini/profile_sys3_anc2_type0.prof</string>
-</dict>
-</plist>
